+++ filter/impl/tps/tps_limit_fix_window_strategy_impl_test.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package impl
+package tps
 
 import (
 	"testing"
